requirements.txt
streamlit==1.38.0
streamlit-aggrid==0.3.5
gspread==6.1.2
google-auth==2.35.0
pandas==2.2.3
streamlit_extras==0.4.7
xlsxwriter==3.2.0
openpyxl==3.1.5
streamlit-calendar
google-api-python-client
st-pages